Refactor using CostCalculators 2
