+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_TEST2.java
@@ -1,13 +1,10 @@
 package mg.mgmap.activity.mgmap.features.routing.profile;
 
-import org.mapsforge.map.datastore.Way;
-
 import mg.mgmap.R;
-import mg.mgmap.generic.graph.WayAttributs;
 
 public class MTB_TEST2 extends GenRoutingProfile {
     public MTB_TEST2( ) {
-        super(new ProfileCostCalculator(8.0, 0.12, 3, 0, -0.27, 2));
+        super(new CostCalculatorForProfile(8.0, 0.12, 3, 0, -0.27, 2));
     }
 
     @Override
