+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_TEST3.java
@@ -6,13 +6,13 @@ import mg.mgmap.generic.graph.WayAttributs;
 public class MTB_TEST3 extends GenRoutingProfile {
 
     public MTB_TEST3( ) {
-        super(new ProfileCostCalculator(8.0, 0.10, 3, 0, -0.27, 2));
+        super(new CostCalculatorForProfile(8.0, 0.10, 3, 0, -0.27, 2));
     }
 
     public void refreshWayAttributes(WayAttributs wayAttributs) {
         if (wayAttributs instanceof WayTagEval ) {
             WayTagEval wayTagEval = (WayTagEval) wayAttributs;
-            wayTagEval.mCostCalculator = new CostCalculatorMTB(wayTagEval, mProfileCostCalculator);
+            wayTagEval.mCostCalculator = new CostCalculatorMTB(wayTagEval, mCostCalculatorForProfile);
         }
     }
     @Override
